pyproject.toml
[project]
name = "attribu-logistics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.6.3",
    "django>=5.2.13",
    "djangorestframework>=3.17.1",
    "gunicorn>=25.3.0",
    "psycopg2-binary>=2.9.12",
    "python-decouple>=3.8",
    "redis>=7.4.0",
    "whitenoise>=6.12.0",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.3.0",
]
